*B* made dist_comp() return a 'dist' object instead of a matrix
conceptually a big change, since dcov.test() tests if the objects are 'dist' and, if not, it calls dist() on them: in this way, the distance is calculated on distances! this is solved by not going through as.matrix() in dist_comp(), i.e. making it return a 'dist' object; this commit also includes all the proper adjustments needed as a result of this change